Partition and replica count in KafkaBinding should be optional (#368)
The Kafka cluster has default partition and replica counts. Users should be able to leverage these defaults if they choose by not explicitly setting the counts in the spec.

This can be particularly useful for replica counts, where different environments can set different default counts to save money, e.g. dev: 1, prod: 3

+++ kafka/src/main/java/io/specmesh/kafka/KafkaApiSpec.java
@@ -98,7 +98,7 @@ public final class KafkaApiSpec {
                                 new NewTopic(
                                                 e.getKey(),
                                                 e.getValue().bindings().kafka().partitions(),
-                                                (short) e.getValue().bindings().kafka().replicas())
+                                                e.getValue().bindings().kafka().replicas())
                                         .configs(e.getValue().bindings().kafka().configs()))
                 .collect(Collectors.toList());
     }
